+++ docs/changes.rst
@@ -4,7 +4,8 @@ Change Log
 
 * 0.2.6 (dev)
 
-  * added preserve_mtime parameter to .put() method, optionally updates the remote file's mtime to match the local file.
+  * added preserve_mtime parameter to .put() method, optionally updates the remote file's st_mtime to match the local file.
+  * added preserve_mtime parameter to .get() method, optionally updates the local file's st_mtime to match the remote file
 
 * 0.2.5 (current, released 2014-05-15)